# Eliminate bounds checks in trinary